Conflicts:

	README
	etc/Makefile.am
	etc/kitcube.ini
	src/kitcube-data/dataserver.cpp
	src/kitcube-devices/Makefile.am
	src/kitcube-devices/ceilometer.h
	src/kitcube-devices/createdevice.cpp
	src/kitcube-devices/daqbinarydevice.cpp
	src/kitcube-devices/daqdevice.cpp
	src/kitcube-devices/daqdevice.h
	src/kitcube-reader/kitcube-reader.cpp
	src/kitcube-reader/reader.cpp
	src/kitcube-reader/reader.h

+++ src/kitcube-devices/ceilometer.cpp
@@ -274,8 +274,8 @@ int Ceilometer::getNumericItem(const char **header, const char *itemTag){
 }
 
 
-unsigned long Ceilometer::getSensorGroup(){
-	unsigned long number;
+unsigned int Ceilometer::getSensorGroup(){
+	unsigned int number;
 	
 	number = 0;
 	buffer = "";
